chore: extract test helpers into shared utils
Ausgelagerte Hilfsfunktionen und Fixtures in core/tests/utils.py und explizite Importe in den Tests.
