+++ core/tests/test_parsing.py
@@ -1,15 +1,45 @@
 import pytest
 from pathlib import Path
+from tempfile import NamedTemporaryFile
 from unittest.mock import patch
 
+from django.core.files.uploadedfile import SimpleUploadedFile
+from django.urls import reverse
+from django.contrib.auth.models import User
+from PIL import Image
+from docx import Document
+
 from .base import NoesisTestCase
-from .test_general import *  # pylint: disable=wildcard-import,unused-wildcard-import
-from ..anlage4_parser import parse_anlage4_dual
-from ..models import Anlage4ParserConfig
-from ..docx_utils import extract_images
-from ..docx_utils import _normalize_header_text
+from ..anlage4_parser import parse_anlage4, parse_anlage4_dual
+from ..docx_utils import (
+    extract_text,
+    get_docx_page_count,
+    get_pdf_page_count,
+    parse_anlage2_table,
+    _normalize_header_text,
+    extract_images,
+)
+from ..models import (
+    BVProject,
+    BVProjectFile,
+    Anlage2Function,
+    Anlage2Config,
+    Anlage2ColumnHeading,
+    Anlage2SubQuestion,
+    Anlage4ParserConfig,
+    Anlage4Config,
+    AntwortErkennungsRegel,
+)
 from ..parsers import ExactParser
-from ..llm_tasks import worker_a4_plausibility as check_anlage4_item_plausibility
+from ..text_parser import parse_anlage2_text
+from ..llm_tasks import (
+    analyse_anlage4,
+    analyse_anlage4_async,
+    worker_anlage4_evaluate,
+    worker_a4_plausibility as check_anlage4_item_plausibility,
+)
+
+pytestmark = pytest.mark.usefixtures("seed_db")
 
 @pytest.mark.usefixtures("prepared_files")
 class DocxExtractTests(NoesisTestCase):
